split models out
